webserver/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
python-whois>=0.9.0
joblib>=1.3.0
gunicorn>=21.0.0
